+++ app/src/main/java/com/raphaelweis/rcube/data/repositories/SolvesRepository.kt
@@ -6,11 +6,11 @@ import kotlinx.coroutines.flow.Flow
 interface SolvesRepository {
     fun getAllSolvesStream(): Flow<List<Solve>>
 
-    fun getSolveStream(id: Int): Flow<Solve?>
+    fun getSolveStream(id: Long): Flow<Solve?>
 
     suspend fun insertSolve(solve: Solve)
 
-    suspend fun deleteSolve(id: Int)
+    suspend fun deleteSolve(id: Long)
 
     suspend fun updateSolve(solve: Solve)
 }
